Kohle-Kartenfilter

+++ SimulateTheWorld.World.Data/Data/Container/TerrainValues.cs
@@ -4,9 +4,9 @@ public class TerrainValues
 {
     public float Height { get; set; }
 
-    public double Coal { get; set; }
-    public double IronOre { get; set; }
-    public double GoldOre { get; set; }
-    public double Oil { get; set; }
-    public double Gas { get; set; }
+    public float Coal { get; set; }
+    public float IronOre { get; set; }
+    public float GoldOre { get; set; }
+    public float Oil { get; set; }
+    public float Gas { get; set; }
 }
